Update class.reader.php
The server was down from 12:30ish to 5:15, so I could not finish testing. I will keep checking it tonight and try to finish testing. All that may be needed is to add more fields in the table to match the data. Also, the field names need to match the definitions in the file. I want to verify this with tests but it seems to be confimed in their documentation.